requirements.txt
langchain-community
PyPDF2
sentence-transformers
qdrant-client
google-generativeai
python-dotenv
fastapi
uvicorn
pydantic
langgraph
pytesseract
psycopg2-binary
unstructured[docx]